feat(controller): Added Catalog controller
