+++ src/main/java/cz/muni/aphos/openapi/CatalogApiController.java
@@ -1,0 +1,16 @@
+package cz.muni.aphos.openapi;
+
+import cz.muni.aphos.openapi.models.Catalog;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/catalog")
+public class CatalogApiController implements CatalogApi{
+    @Override
+    public ResponseEntity<String[]> getCatalogs() {
+        return new ResponseEntity<>(Catalog.getCatalogs(), HttpStatus.OK);
+    }
+}
