+++ src/main/java/cz/muni/aphos/openapi/UserApi.java
@@ -45,19 +45,5 @@ public interface UserApi {
     ResponseEntity<User> getUserByUsername(@NotBlank @Parameter(in = ParameterIn.QUERY, required=true)
                                            @Valid @RequestParam(value = "username") String username);
 
-    @Operation(summary = "Find all catalogs", description = "Returns catalogs", tags={ "Catalog" })
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Successful operation",
-                    content = @Content(array = @ArraySchema(schema = @Schema(implementation = Catalog.class)))),
-
-
-            @ApiResponse(responseCode = "400", description = "Error",
-                    content = @Content(schema = @Schema(implementation = ErrorMessage.class)))
-    })
-    @RequestMapping(value = "/getCatalogs",
-            produces = { "application/json", "application/xml" },
-            method = RequestMethod.GET)
-    ResponseEntity<String[]> getCatalogs();
-
 }
