+++ src/main/java/cz/muni/aphos/openapi/UserApiController.java
@@ -68,13 +68,6 @@ public class UserApiController implements UserApi {
         return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
-    /**
-     * @return
-     */
-    @Override
-    public ResponseEntity<String[]> getCatalogs() {
-        return new ResponseEntity<>(Catalog.getCatalogs(),HttpStatus.OK);
-    }
 
 
 }
